lab2/backend/requirements.txt
django>=5.0.0
djangorestframework>=3.14.0
markdown>=3.5.0
django-filter>=23.5
django-cors-headers>=4.3.1
psycopg[binary,pool]>=3.1.12
faker>=20.0.0
gunicorn>=21.2.0

